Test productor-consumidor
Subo un test de productor consumidor hecho con lo semáforos y qué se yo. 

El test recién lo hice andar y no anda bien:
 - Lee basuras.
 - Un ctrl+C deja el ipcs lleno de bostas.
 - etc...
 
Pero... tiene algo interesante el test que es ya de una envergadura más importante que los anteriores. Y la cuestión es que por ejemplo me topé con el problema de dónde inicializar los recursos compartidos. Decidí hacerlo en el proceso padre, al principio, pero de una forma extraña. Si iniciaba a los recursos al principio y les decía a esos objetos "ustedes son los dueños de los recursos", y etos objetos morían al final del proceso, cuando este se fork()eara, los hijos tendrían copias de estos objetos, copias que tienen indicadas que son las dueñas de los recursos, pero no lo son! 

Lo que hice es que el padre crea objetos que inicializan las cosas, pero luego los destruye, "avisándoles" a los objetos que estos no deben liberar los recursos antes de destruirlos. Esto deja los recursos ya inicializados para los próximos procesos hijos, y también deja al padre sin objetos locales "raros"...

Una vez hecho esto, cuando los hijos usan los recursos, ya no los utilizan como dueños, por ende no los destruyen cuando termian. 

Y Finalmente, para liberar los recursos, el padre, luego de haber esperado a todos los hijos (que ahora tienen el problema que no mueren nunca xD) vuelve a crear esos objetos que reclaman los recursos al ser destruidos.

Bueno... la cuestión es que para que esto tuviese algo de sentido, agregué una clase base Resource, de la que hice extender a la mayoría de las cosas (excepto message queue porque me dio paja jeje). Resource tiene, por ahora, nada más los métodos para consultar y setear el atributo "ownResources" que le indica al Resource si este es "dueño" de los recursos externos o no... Pero por ahora son los hijos los que consultan esta cosa y deciden qué hacer.

En fin.. eso es todo..

+++ trunk/concu-chat/src/ipc/fifo.cpp
@@ -24,7 +24,9 @@ bool isFifoFile(const string& pathName)
 }
 
 
-Fifo::Fifo(const string& pathName) : _pathName(pathName)
+Fifo::Fifo(const string& pathName, bool ownResource):
+		Resource(ownResource),
+		_pathName(pathName)
 {
 	bool mknodEror = mknod(pathName.c_str(), S_IFIFO | 0666, 0) == -1;
 
@@ -44,12 +46,15 @@ Fifo::Fifo(const string& pathName) : _pathName(pathName)
 }
 
 
-Fifo::~Fifo()
+Fifo::~Fifo() throw ()
 {
-	bool unlinkError = unlink(_pathName.c_str()) == -1;
+	if (ownResources())
+	{
+		bool unlinkError = unlink(_pathName.c_str()) == -1;
 
-	if (unlinkError && errno != ENOENT) // ENOENT = No such file.
-		perror("~Fifo(): Could not unlink the fifo's file");
+		if (unlinkError && errno != ENOENT) // ENOENT = No such file.
+			perror("~Fifo(): Could not unlink the fifo's file");
+	}
 }
